[gaim-migrate @ 7981] Code cleanup to get rid of all warnings for AMD64 from augustus at linuxhardware dot org
--HG--
extra : convert_revision : afd74f46cee2f4a4c349818ea19d3d1d33e90777

+++ src/gtknotify.c
@@ -206,8 +206,8 @@ gaim_gtk_notify_emails(size_t count, gboolean detailed,
 
 	/* Descriptive label */
 	detail_text = g_strdup_printf(
-		ngettext("%s has %d new message.", "%s has %d new messages.", count),
-		*tos, count);
+		ngettext("%s has %d new message.", "%s has %d new messages.", (int) count),
+		*tos, (int) count);
 
 	if (count == 1) {
 		char *from_text = NULL, *subject_text = NULL;
